SAAS/requirements.txt
asgiref==3.7.2
Django==3.2.19
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap3==23.1
django-bootstrap3-datetimepicker-3==2.6.0
django-ninja==0.22.1
django-mptt==0.11.0
importlib-metadata==2.1.3
pip==18.1
pydantic==1.10.8
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
zipp==3.15.0
django-mathfilters
django-cleanup
django-reversion==5.0.4
openpyxl
pandas==0.24.2
xlrd==1.2.0
XlsxWriter==1.3.7
image==1.5.33
qrcode==6.1
arrow==0.14.1
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.8.5